1. reorganized the code and deleted redundant functions
